Bug 26129: Add Koha::Configuration(s) classes
This patch adds the Koha::Configuration(s) classes and tests. It is
similar to the Koha::CirculationRules(s) classes, but with simpler use
cases for now. There's no special handling for '*' as it is expected to
be passed undef instead.

It allows to set arbitrary configuration entries either globally or for
specific library_id, category_id or item_type.

Attribute names match the ones agreed for the API.

To test:
1. Apply this patches
2. Run:
   $ updatedatabase
=> SUCCESS: The atomic update is applied correctly
3. Run:
   $ reset_all
=> SUCCESS: kohastructure.sql is loaded without issues
4. Run:
   $ kshell
  k$ prove t/db_dependent/Koha/Configurations.t
=> SUCCESS: Tests pass!
5. Sign off :-D

Sponsored-by: ByWater Solutions

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: David Nind <david@davidnind.com>