Bug 38924: Add Koha::Patron::Quota::Usage(s) classes
Add ::Usage(s) classes for working with quota allocation usage.

We overload store to ensure data consistency for the checkout_id field
which is linked to either the issue or old_issue table appropriatly.

We include a checkout relationship that handles the lack of foreign key
due to issues/old_issues.